requirements.txt
anthropic
python-telegram-bot
python-dotenv
httpx[socks]